requirements.txt
matplotlib==3.7.5
numpy==1.24.3
opencv-python==4.12.0.88
Pillow==10.2.0
plyfile==1.0.3
poselib==2.0.4
PyYAML==5.4.1
scikit-learn==1.3.2
scipy==1.10.1
trimesh==4.7.3
faiss-cpu==1.8.0
yacs==0.1.8
loguru==0.7.3
pytorch-lightning==1.3.5
torchmetrics==0.6.0
einops==0.8.1
mamba-ssm==2.2.2
kornia==0.7.0
timm==1.0.15
thop==0.1.1.post2209072238
tqdm==4.67.1
